rename cwlogsprovision extension to cwlogsauth extension
