+++ extension/cwlogsauth/config.go
@@ -1,14 +1,14 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT
 
-package cwlogsprovision
+package cwlogsauth
 
 import "go.opentelemetry.io/collector/component"
 
-// Config for the cwlogsprovision extension.
+// Config for the cwlogsauth extension.
 type Config struct {
 	// Auth is the ID of an auth extension (e.g., sigv4auth) to chain with.
-	// The cwlogsprovision extension wraps this auth's RoundTripper to add
+	// The cwlogsauth extension wraps this auth's RoundTripper to add
 	// lazy log group/stream creation before forwarding requests.
 	Auth *component.ID `mapstructure:"auth,omitempty"`
 }
