+++ extension/cwlogsauth/extension.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT
 
-package cwlogsprovision
+package cwlogsauth
 
 import (
 	"context"
