+++ service/defaultcomponents/components.go
@@ -66,7 +66,7 @@ import (
 	"go.opentelemetry.io/collector/receiver/otlpreceiver"
 
 	"github.com/aws/amazon-cloudwatch-agent/extension/agenthealth"
-	"github.com/aws/amazon-cloudwatch-agent/extension/cwlogsprovision"
+	"github.com/aws/amazon-cloudwatch-agent/extension/cwlogsauth"
 	"github.com/aws/amazon-cloudwatch-agent/extension/entitystore"
 	"github.com/aws/amazon-cloudwatch-agent/extension/k8smetadata"
 	"github.com/aws/amazon-cloudwatch-agent/extension/server"
@@ -162,7 +162,7 @@ func Factories() (otelcol.Factories, error) {
 	if factories.Extensions, err = otelcol.MakeFactoryMap[extension.Factory](
 		agenthealth.NewFactory(),
 		awsproxy.NewFactory(),
-		cwlogsprovision.NewFactory(),
+		cwlogsauth.NewFactory(),
 		entitystore.NewFactory(),
 		k8smetadata.NewFactory(),
 		server.NewFactory(),
